add credential renamed
