+++ src/components/PasswordManager.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import useSWR, { mutate } from "swr";
 import { fetchAll, post, remove } from "../data/apiClient";
 import { authService } from "../services/authService";
-import AddPwd from "./AddPwd";
+import AddPwd from "./AddCredential";
 import { CredentialEntry } from "./Credential";
 import CredentialList from "./CredentialList";
 import { ChevronLeftIcon, PlusIcon } from "./icons/Icons";
